+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 * Set proxy environment variables if preset in config. (CHEF-4712)
 * Automatically enable verify_api_cert when running chef-client in local-mode. (Chef Issues 1464)
 * Add helper to warn for broken [windows] paths. (CHEF-5322)
+* Send md5 checksummed data for registry key if data type is binary, dword, or qword. (Chef-5323)
 
 ## Release: 11.12.4 (04/30/2014)
 http://www.getchef.com/blog/2014/04/30/release-chef-client-11-12-4-ohai-7-0-4/